Break manater_test.go to reusable parts
