+++ core/connection/stub_promises.go
@@ -1,0 +1,35 @@
+/*
+ * Copyright (C) 2018 The "MysteriumNetwork/node" Authors.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package connection
+
+import "github.com/mysteriumnetwork/node/service_discovery/dto"
+
+type fakePromiseIssuer struct {
+	startCalled bool
+	stopCalled  bool
+}
+
+func (issuer *fakePromiseIssuer) Start(proposal dto.ServiceProposal) error {
+	issuer.startCalled = true
+	return nil
+}
+
+func (issuer *fakePromiseIssuer) Stop() error {
+	issuer.stopCalled = true
+	return nil
+}
