+++ core/connection/stub_stats.go
@@ -1,0 +1,47 @@
+/*
+ * Copyright (C) 2018 The "MysteriumNetwork/node" Authors.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package connection
+
+import (
+	"time"
+
+	"github.com/mysteriumnetwork/node/client/stats"
+)
+
+type fakeSessionStatsKeeper struct {
+	sessionStartMarked, sessionEndMarked bool
+}
+
+func (fsk *fakeSessionStatsKeeper) Save(stats stats.SessionStats) {
+}
+
+func (fsk *fakeSessionStatsKeeper) Retrieve() stats.SessionStats {
+	return stats.SessionStats{}
+}
+
+func (fsk *fakeSessionStatsKeeper) MarkSessionStart() {
+	fsk.sessionStartMarked = true
+}
+
+func (fsk *fakeSessionStatsKeeper) GetSessionDuration() time.Duration {
+	return time.Duration(0)
+}
+
+func (fsk *fakeSessionStatsKeeper) MarkSessionEnd() {
+	fsk.sessionEndMarked = true
+}
